requirements.txt
boto3==1.10.48
botocore==1.13.48
docopt==0.6.2
docutils==0.15.2
jmespath==0.9.4
lark-parser==0.7.8
parliament==0.4.12
python-dateutil==2.8.1
python-hcl2==0.2.2
PyYAML==5.3
s3transfer==0.2.1
six==1.13.0
urllib3==1.25.7
